pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "resume-env"
version = "1.0.0"
description = "ResumeEnv: Job Application Optimizer — OpenEnv environment"
requires-python = ">=3.10"
dependencies = [
    "openenv-core>=0.2.0",
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.29.0",
    "pydantic>=2.0.0",
    "openai>=1.0.0",       # for inference.py (also works with any OpenAI-compatible API)
    "httpx>=0.27.0",
    "websockets>=12.0",
]

[project.scripts]
server = "server.app:main"

[tool.hatch.build.targets.wheel]
packages = ["."]
